requirements.txt
flask>=3.0.0
streamlit>=1.28.0
PyMuPDF>=1.23.0
openai>=1.0.0
pillow>=10.0.0
imagehash>=4.3.0
python-dotenv>=1.0.0
requests>=2.31.0

